+++ src/_internal/operators/inc.ts
@@ -1,20 +1,21 @@
-import { ArrayOrObject, Callback, RawObject } from "mingo/types";
-import { walk } from "mingo/util";
+import { ArrayOrObject, RawObject } from "mingo/types";
 
 import { UpdateOptions } from "../types";
+import { Action, applyUpdate, walkExpression } from "../util";
 
 /** Increments a field by a specified value. */
 export const $inc = (
   obj: RawObject,
   expr: Record<string, number>,
+  arrayFilters: RawObject[],
   options: UpdateOptions
 ) => {
-  for (const [selector, val] of Object.entries(expr)) {
+  walkExpression(expr, arrayFilters, ((val, node, queries) => {
     let changed = false;
-    walk(obj, selector, ((o: ArrayOrObject, k: string) => {
+    applyUpdate(obj, node, queries, (o: ArrayOrObject, k: number) => {
       o[k] = (o[k] as number) + val;
       changed = true;
-    }) as Callback);
-    if (changed) options.emit(selector);
-  }
+    });
+    if (changed) options.emit(node.parent);
+  }) as Action<number>);
 };