+++ src/_internal/operators/pullAll.ts
@@ -7,11 +7,12 @@ import { $pull } from "./pull";
 export const $pullAll = (
   obj: RawObject,
   expr: Record<string, RawArray>,
+  arrayFilters: RawObject[],
   options: UpdateOptions
 ) => {
   const pullExpr: Record<string, RawObject> = {};
   Object.entries(expr).forEach(([k, v]) => {
     pullExpr[k] = { $in: v };
   });
-  $pull(obj, pullExpr, options);
+  $pull(obj, pullExpr, arrayFilters, options);
 };