+++ src/_internal/operators/set.ts
@@ -1,25 +1,30 @@
 import { AnyVal, ArrayOrObject, Callback, RawObject } from "mingo/types";
-import { cloneDeep, isEqual, walk } from "mingo/util";
+import { cloneDeep, isEqual } from "mingo/util";
 
 import { UpdateOptions } from "../types";
+import { applyUpdate, walkExpression } from "../util";
 
 /** Replaces the value of a field with the specified value. */
 export const $set = (
   obj: RawObject,
   expr: Record<string, AnyVal>,
+  arrayFilters: RawObject[],
   options: UpdateOptions
 ) => {
-  for (const [selector, val] of Object.entries(expr)) {
-    walk(
+  walkExpression(expr, arrayFilters, (val, node, queries) => {
+    let changed = false;
+    applyUpdate(
       obj,
-      selector,
-      ((o: ArrayOrObject, k: string | number) => {
+      node,
+      queries,
+      ((o: ArrayOrObject, k: string) => {
         if (!isEqual(o[k], val)) {
           o[k] = cloneDeep(val);
-          options.emit(selector);
+          changed = true;
         }
       }) as Callback,
       { buildGraph: true }
     );
-  }
+    if (changed) options.emit(node.parent);
+  });
 };