+++ src/_internal/operators/unset.ts
@@ -1,21 +1,24 @@
-import { Callback, RawObject } from "mingo/types";
-import { has, walk } from "mingo/util";
+import { RawObject } from "mingo/types";
+import { has } from "mingo/util";
 
 import { UpdateOptions } from "../types";
+import { applyUpdate, walkExpression } from "../util";
 
 /** Deletes a particular field */
 export const $unset = (
   obj: RawObject,
   expr: Record<string, "">,
+  arrayFilters: RawObject[],
   options: UpdateOptions
 ) => {
-  for (const selector of Object.keys(expr)) {
-    walk(obj, selector, ((o: RawObject, k: string) => {
-      const changed = has(o, k);
-      if (changed) {
+  walkExpression(expr, arrayFilters, (_, node, queries) => {
+    let changed = false;
+    applyUpdate(obj, node, queries, (o: RawObject, k: string) => {
+      if (has(o, k)) {
         delete o[k];
-        options.emit(selector);
+        changed = true;
       }
-    }) as Callback);
-  }
+    });
+    if (changed) options.emit(node.parent);
+  });
 };