+++ test/operators/addToSet.test.ts
@@ -4,7 +4,7 @@ import { UPDATE_OPTIONS } from "../helper";
 describe("operators/addToSet", () => {
   it("Value to Add is An Array", () => {
     const state = { _id: 1, letters: ["a", "b"] };
-    $addToSet(state, { letters: ["c", "d"] }, UPDATE_OPTIONS);
+    $addToSet(state, { letters: ["c", "d"] }, [], UPDATE_OPTIONS);
     expect(state).toEqual({ _id: 1, letters: ["a", "b", ["c", "d"]] });
   });
 
@@ -14,8 +14,8 @@ describe("operators/addToSet", () => {
       item: "polarizing_filter",
       tags: ["electronics", "camera"]
     };
-    $addToSet(state, { tags: "accessories" }, UPDATE_OPTIONS);
-    $addToSet(state, { tags: "camera" }, UPDATE_OPTIONS);
+    $addToSet(state, { tags: "accessories" }, [], UPDATE_OPTIONS);
+    $addToSet(state, { tags: "camera" }, [], UPDATE_OPTIONS);
     expect(state).toEqual({
       _id: 1,
       item: "polarizing_filter",
@@ -28,6 +28,7 @@ describe("operators/addToSet", () => {
     $addToSet(
       state,
       { tags: { $each: ["camera", "electronics", "accessories"] } },
+      [],
       UPDATE_OPTIONS
     );
     expect(state).toEqual({